# Try always multiplying director tensors.