--- a/source/heating_model/shear_heating.cc
+++ b/source/heating_model/shear_heating.cc
@@ -42,11 +42,7 @@ namespace aspect
       for (unsigned int q=0; q<heating_model_outputs.heating_source_terms.size(); ++q)
         {
           const SymmetricTensor<4,dim> &C = material_model_outputs.stress_strain_directors[q];
-          const SymmetricTensor<2,dim> &directed_strain_rate = ((C == dealii::identity_tensor<dim> ())
-                                                                ?
-                                                                C * material_model_inputs.strain_rate[q]
-                                                                :
-                                                                material_model_inputs.strain_rate[q]);
+          const SymmetricTensor<2,dim> &directed_strain_rate = C * material_model_inputs.strain_rate[q];
 
           const SymmetricTensor<2,dim> stress =
             2 * material_model_outputs.viscosities[q] *