For speed comparison with #327