@jperryhouts 2015-10-27
Um. This is somehow faster that the comparison method now. I have no idea what to think of that... And my test case is failing, so maybe I just broke it somehow? :/ I'll play with it more tomorrow.