@jperryhouts 2015-10-27
Oh, the test was failing because of a bug in the anisotropic shear heating, fixed by #655 

It looks like the times (for the ctest models) seem to be comparable between the two methods (random variations, probably just because of other system load, make some difference).

However, the "always-multiply" method slows down by a factor of ~2 relative to the "comparison" method in Debug mode... For some tests, but not all.