@jperryhouts 2015-10-27
For example: In Debug mode, with "compare" method:
```
  1/264 Test   #1: adiabatic_conditions ...........................................   Passed   23.59 sec
  2/264 Test   #2: adiabatic_heating ..............................................   Passed   23.29 sec
  3/264 Test   #3: adiabatic_heating_simplified ...................................   Passed   22.91 sec
  4/264 Test   #4: adiabatic_initial_conditions ...................................   Passed   23.57 sec
  5/264 Test   #5: adiabatic_initial_conditions_chunk .............................***Failed    7.75 sec
  6/264 Test   #6: adiabatic_initial_conditions_chunk_3d ..........................***Failed  154.42 sec
  7/264 Test   #7: adiabatic_initial_conditions_constant ..........................   Passed   23.99 sec
  8/264 Test   #8: adiabatic_initial_conditions_subadiabaticity ...................   Passed   23.33 sec
  9/264 Test   #9: adiabatic_initial_conditions_without_boundary ..................   Passed    6.57 sec
 10/264 Test  #10: advect_mesh_vertically .........................................***Failed   24.83 sec
 11/264 Test  #11: always_refine ..................................................   Passed    6.34 sec
 12/264 Test  #12: anisotropic_viscosity ..........................................   Passed   16.24 sec
 13/264 Test  #13: artificial_postprocess .........................................   Passed   12.06 sec
...
```
And with "multiply" method:
```
  1/264 Test   #1: adiabatic_conditions ...........................................   Passed   19.01 sec
  2/264 Test   #2: adiabatic_heating ..............................................   Passed   11.88 sec
  3/264 Test   #3: adiabatic_heating_simplified ...................................   Passed   11.80 sec
  4/264 Test   #4: adiabatic_initial_conditions ...................................   Passed   18.94 sec
  5/264 Test   #5: adiabatic_initial_conditions_chunk .............................***Failed    6.60 sec
  6/264 Test   #6: adiabatic_initial_conditions_chunk_3d ..........................***Failed   78.58 sec
  7/264 Test   #7: adiabatic_initial_conditions_constant ..........................   Passed   19.06 sec
  8/264 Test   #8: adiabatic_initial_conditions_subadiabaticity ...................   Passed   18.89 sec
  9/264 Test   #9: adiabatic_initial_conditions_without_boundary ..................   Passed    5.29 sec
 10/264 Test  #10: advect_mesh_vertically .........................................***Failed   15.15 sec
 11/264 Test  #11: always_refine ..................................................   Passed    5.36 sec
 12/264 Test  #12: anisotropic_viscosity ..........................................   Passed    8.49 sec
 13/264 Test  #13: artificial_postprocess .........................................   Passed    4.74 sec
...
```