@jperryhouts 2015-10-27
I ran all the tests for both the debug and release mode in each of the cases (the comparison version, the multiplication version, and the original version: with no director tensors at all). Here are the timing results:
```
debug-multiply.tests:      Total Test time (real) = 1247.56 sec
debug-nodirection.tests:   Total Test time (real) = 791.78 sec
debug-compare.tests:       Total Test time (real) = 788.17 sec
release-multiply.tests:    Total Test time (real) = 497.04 sec
release-nodirection.tests: Total Test time (real) = 484.43 sec
release-compare.tests:     Total Test time (real) = 487.22 sec
```