@bangerth 2015-11-11
I finally found some time and ran a modification of shell-simple-3d on my system, and here is what I have:

Current master, debug mode: 
```
| Assemble Stokes system          |         8 |       121s |        28% |
| Assemble temperature system     |         8 |       156s |        36% |
| Build Stokes preconditioner     |         3 |      47.2s |        11% |
| Build temperature preconditioner|         8 |     0.551s |      0.13% |
| Solve Stokes system             |         8 |      38.2s |       8.9% |
```

Current master, release mode:
```
| Assemble Stokes system          |         8 |      5.96s |         7% |
| Assemble temperature system     |         8 |      5.38s |       6.3% |
| Build Stokes preconditioner     |         3 |      6.38s |       7.5% |
| Build temperature preconditioner|         8 |     0.564s |      0.66% |
| Solve Stokes system             |         8 |      62.6s |        74% |
```


With your patch, debug mode:
```
| Assemble Stokes system          |         8 |       811s |        63% |
| Assemble temperature system     |         8 |       160s |        12% |
| Build Stokes preconditioner     |         3 |       199s |        15% |
| Build temperature preconditioner|         8 |     0.565s |     0.044% |
| Solve Stokes system             |         8 |      50.3s |       3.9% |
```

With your patch, release mode:
```
| Assemble Stokes system          |         8 |      22.2s |        23% |
| Assemble temperature system     |         8 |      5.16s |       5.4% |
| Build Stokes preconditioner     |         3 |      8.31s |       8.6% |
| Build temperature preconditioner|         8 |     0.542s |      0.56% |
| Solve Stokes system             |         8 |      55.9s |        58% |
```

It's not quite clear to me what is going on with the different solver times (maybe something else was running on the system), but in either case, the new assembly is incurring a sizable performance hit, even in release mode. This is with the current version of the pull request.

This was in 3d, where I expect the difference to be largest. I suspect that the tests mostly run in 2d, so maybe the difference isn't so visible there.

Would you mind repeating your tests with a 3d testcase and with all 6 versions you had? I used `cookbook/shell_simple_3d.prm` with the following changes:
```
diff --git a/cookbooks/shell_simple_3d.prm b/cookbooks/shell_simple_3d.prm
index 1a26209..fbad434 100644
--- a/cookbooks/shell_simple_3d.prm
+++ b/cookbooks/shell_simple_3d.prm
@@ -4,7 +4,7 @@
 
 set Dimension                              = 3
 set Use years in output instead of seconds = true
-set End time                               = 1.5e9
+set End time                               = 1.5e6
 set Output directory                       = output
 
 
@@ -62,15 +62,15 @@ end
 
 
 subsection Mesh refinement
-  set Initial global refinement          = 2
-  set Initial adaptive refinement        = 3
+  set Initial global refinement          = 1
+  set Initial adaptive refinement        = 2
   set Strategy                           = temperature
   set Time steps between mesh refinement = 15
 end
 
 
 subsection Postprocess
-  set List of postprocessors = visualization, velocity statistics, temperature statistics, heat flux statistics, depth average
+  set List of postprocessors =
 
   subsection Visualization
     set Output format                 = vtu
```
With 8 processors, this should run in a few minutes (around 7 minutes in debug mode, about 1.5 minutes in release mode).